XSSAuditor takes post body from current request, not the original request.
In the face of a redirect, the information in the original body can't be
reflected in the final page, when we redirect from post to get, since the
get has no body.  And for a 307-style redirect from post to post, the body
will appear in the final post.

This avoids some false positives and also the possibility of some info
leaks from the original post.

BUG=331725
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/128823003

git-svn-id: svn://svn.chromium.org/blink/trunk@164749 bbb929c8-8fbe-4397-9dbb-9b2b20218538