Remove unnecessary DrawingBuffer::m_lastColorBuffer.
DrawingBuffer::m_lastColorBuffer has the same role to m_frontColorBuffer, which
is the reference of the front buffer.

In addition, m_lastColorBuffer is used only when the drawing buffer is preserved. It
makes hard to understand code.

So this CL removes DrawingBuffer::m_lastColorBuffer.

This CL does not change logic. To ensure it, new test is added.
TEST=fast/canvas/webgl/draw-webgl-to-canvas-2d.html

BUG=331181

Review URL: https://codereview.chromium.org/115273004

git-svn-id: svn://svn.chromium.org/blink/trunk@164809 bbb929c8-8fbe-4397-9dbb-9b2b20218538