Include compositing reasons in GraphicsLayerDebugInfo
Previously, this included only layout rects. We should now be able to stuff
arbitrary debug info in there, and with this CL we add the compositing reasons.

There are 3 copies of the CompositingReasons enum. We have CompositingReasons.h
WebCompositingReasons.h and compositing_reasons.h (in cc). The only reason for
the duplication is that we plumbed the reasons all the way from blink into cc
so that they could be converted to strings for tracing. Now that we include
the compositing reasons in the debug info this is unnecessary. I've tried to
get all blink code using CompositingReasons.h, but to avoid layering violations
this required putting the file in platform/graphics. Admittedly this is a bit
of a bummer and I'm open to suggestions for a better home.

Note, this change depends on https://codereview.chromium.org/128303003/
BUG=None

Review URL: https://codereview.chromium.org/129363002

git-svn-id: svn://svn.chromium.org/blink/trunk@164814 bbb929c8-8fbe-4397-9dbb-9b2b20218538