Teach mapFill{X,Y}Position about initial values to individual background layers
In <https://codereview.chromium.org/35303002>, it was falsely assumed
that the initial values for background-position-{x,y} only came from
the property as a whole, ignoring that the initial value must be taken
into account *for each layer* too. Otherwise we will leak background
positions from one layer to the next, causing the positioning bug the
reporter found.

This is a partial revert of r160243 along with an added js test.

BUG=327307

Review URL: https://codereview.chromium.org/128203003

git-svn-id: svn://svn.chromium.org/blink/trunk@164877 bbb929c8-8fbe-4397-9dbb-9b2b20218538