[FastTextAutosizer] Fix nested inline lineheight bugs
This patch is a followup to [1] which failed to account for
nested inlines. With this patch we now correctly inflate the
parent of text nodes which fixes some lineheight issues when
the parent controls the child's lineheight. See
TextAutosizer::processContainer which has the exact same logic.

With this patch we pass an additional 8 tests!

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=163904

BUG=

Review URL: https://codereview.chromium.org/133153002

git-svn-id: svn://svn.chromium.org/blink/trunk@164930 bbb929c8-8fbe-4397-9dbb-9b2b20218538