Remove MarkupAccumulator::appendNodeValue() method
Remove MarkupAccumulator::appendNodeValue() method and merge it into
MarkupAccumulator::appendText() as this is the only caller. Now that we know
the Node is actually a Text node, update the implementation to call
CharacterData::data() instead of Node::nodeValue() which is virtual.

Also update CharacterData::data() to return a const reference to avoid
unnecessary ref churn on the String.

R=esprehn, abarth

Review URL: https://codereview.chromium.org/135443003

git-svn-id: svn://svn.chromium.org/blink/trunk@165154 bbb929c8-8fbe-4397-9dbb-9b2b20218538