Refactor calculation of hasRx and hasRy values in SVGPathData
This is a merge of http://trac.webkit.org/changeset/162537 by Peter
Molnar <pmolnar.u-szeged@partner.samsung.com>.

This patch is a followup to a Blink patch by Mrunal [1] which I
reviewed. In merging that patch to WebKit, Darin pointed out the code
could be cleaner [2] and I'm making that change back in Blink as well.
The actual code change is to not read rx and ry twice.

[1] https://src.chromium.org/viewvc/blink?revision=152376&view=revision
[2] https://bugs.webkit.org/show_bug.cgi?id=127337#c4

No change in behavior as this is just a trivial refactoring.

BUG=336897

Review URL: https://codereview.chromium.org/143413003

git-svn-id: svn://svn.chromium.org/blink/trunk@165574 bbb929c8-8fbe-4397-9dbb-9b2b20218538