Use WeakPtr<> for FormAssociatedElement::m_form and HTMLImageElement::m_form.
We found we didn't need to do anything other than clearing m_form in
disassociation process of ~HTMLFormElement. So we can use WeakPtr<>.
Also, this adds WeakPtr::operator->.

* What we did, and why are they unnecessary?
 FormAssociatedElement::formWillBeDestroyed did the follwoings:
 A) Called FormAssociatedElement::willChangeForm
 B) Called FormAssociatedElement::didChangeForm with m_form=0.

A was actually HTMLInputElement::removeFromRadioButtonGroup. This is unnecessary
because the radio button group is owned by the HTMLFormElement, and
HTMLInputElement doesn't have any references to the radio button group.

B was actually HTMLInputElement::addToRadioButtonGroup. If the input element is
not inDocument state, this function does nothing. Otherwise, the Document is
dying and we don't need to update radio button groups of Document.

BUG=334556
TEST=none; no user-visible behavior changes.

Review URL: https://codereview.chromium.org/136173006

git-svn-id: svn://svn.chromium.org/blink/trunk@165645 bbb929c8-8fbe-4397-9dbb-9b2b20218538