Make direction explicit for UChar/LChar version of constructTextRun
Explicitly pass the text direction to the UChar and LChar versions of
constructTextRun as well as the RenderText wrapper. This allows callers
to specify the desired direction and makes cases where it is hard-coded
more obvious. Part of a larger set of changes to make all callers
specify the desired direction.

BUG=333004
R=dglazkov@chromium.org

Review URL: https://codereview.chromium.org/132393004

git-svn-id: svn://svn.chromium.org/blink/trunk@165649 bbb929c8-8fbe-4397-9dbb-9b2b20218538